requirements.txt
flask
mysql-connector-python
bcrypt
python-dotenv
